+++ CHANGES.md
@@ -38,3 +38,5 @@ Please follow the [`CHANGES.md` modification guidelines](https://github.com/dcos
 * Updated Exhibitor to version running atop [Jetty 9.4.30](https://github.com/dcos/exhibitor/commit/e6e232e1)
 
 * Ensure Docker network for Calico is eventually created correctly following failures. (D2IQ-70674)
+
+* Check that `spartan` ips (`198.51.100.1-3`) are not listed as upstream resolvers. (COPS-4616)